app/requirements.txt
pillow==10.3.0

numpy~=1.26.4